feat/MSSDK-2126: changes from CR
